altered start logic an made standard plugins self healing. This will catch installtions pre 3.3-3 too, because they will get a new subvolume for plugins and therefore cas and activerecord will have to be reinstalled

+++ containers/redmine/startup.sh
@@ -68,22 +68,27 @@ else
   # Remove default admin account
   sql "DELETE FROM users WHERE login='admin';"
 
-  # Install Redmine CAS plugin
-  echo "installing CAS plugin"
-  curl -L -o casplugin${RMCASPLUGINVERSION}.tar.gz https://github.com/cloudogu/redmine_cas/archive/${RMCASPLUGINVERSION}.tar.gz \
-    && tar -xzf casplugin${RMCASPLUGINVERSION}.tar.gz \
-    && mv redmine_cas-${RMCASPLUGINVERSION} ${WORKDIR}/plugins/redmine_cas \
-    && rm casplugin${RMCASPLUGINVERSION}.tar.gz
-  
+fi
+
+# Install base plugins for cloudogu
+if [ $(ls -l ${WORKDIR}/plugins/ | grep "redmine_activerecord" | wc -l) -eq 0 ]; then
   # Install redmine_activerecord_session_store plugin to make Single Sign-Out work
   echo "installing redmine_activerecord_session_store plugin"
   curl -L -o redmine_AR_session_store.tar.gz https://github.com/pencil/redmine_activerecord_session_store/archive/v0.0.1.tar.gz \
     && tar -xzf redmine_AR_session_store.tar.gz \
     && mv redmine_activerecord_session_store-0.0.1 ${WORKDIR}/plugins/redmine_activerecord_session_store \
     && rm redmine_AR_session_store.tar.gz
-
+fi
+if [ $(ls -l ${WORKDIR}/plugins/ | grep "redmine_cas" | wc -l) -eq 0 ]; then
+  # Install Redmine CAS plugin
+  echo "installing CAS plugin"
+  curl -L -o casplugin${RMCASPLUGINVERSION}.tar.gz https://github.com/cloudogu/redmine_cas/archive/${RMCASPLUGINVERSION}.tar.gz \
+    && tar -xzf casplugin${RMCASPLUGINVERSION}.tar.gz \
+    && mv redmine_cas-${RMCASPLUGINVERSION} ${WORKDIR}/plugins/redmine_cas \
+    && rm casplugin${RMCASPLUGINVERSION}.tar.gz
 fi
 
+# Install Plugins
 echo "running plugins migrations..."
   rake redmine:plugins:migrate RAILS_ENV=$RAILS_ENV -f ${WORKDIR}/Rakefile
 echo "plugins migrations... done"
